Buffer hash input (#29)
This refactors how the hash algorithm is used when computing `stable_hash` so that
I can optimize performance more easily. Then it implements a wrappers that allow
data to be stored in a buffer before the hash is computed:

BufferedHashState: This stores data and "flushes" it when the buffer gets big enough
by computing a hash on the buffer data. It handles nesting by using a delimiter buffer that is hashed after the content being hashed.

Note that I plan to make a few more changes that will again change the hash value (#30), so this does not bump the version in Project.toml since there is a planned follow-up PR that will be merged before releasing a new version.