fix Base.Fix1/2 error (#34)
HashVersion{V} for V > 1 used to crash for types of Base.Fix1 and Base.Fix2. This adds hash_method methods to address that.

This involved generating a new reference tests for HashVersion{1} with reference files for before implementing my fix, to ensure that the new changes don't alter the hash value for that context.

This change did not noticeably affect the benchmarks.