Make Julia 1.10 type ids stable (#43)
Julia 1.10 changes the string representation of NamedTuple types, and we use these string representations to generate stable hash values for types:

What was once:

```
Base.NamedTuple{(:a,:b),Tuple{Int64,NamedTuple{(:x,:y),Tuple{Int64,Int64}}}}
```

Is now: 

```
@NamedTuple{a::Int64, b::@NamedTuple{x::Int64, y::Int64}}
```

As I see it we have two options for how to handle things moving forward with 1.10: we could start to create a function that generates a string (or some other stable representation) from the internal representation of a type OR we can parse the string generated in Julia 1.10 to generate the string of a type from Julia 1.9 and lower. The former is a long-term solution, but it changes existing hashes (and I have a sketch of a plan for that below in the comments). I think the latter needs to happen so we can avoid breaking existing hashes.

So in this PR I am going with the latter option, using PikaParser to parse out the bits of syntax I need to re-arrange named tuple strings into the pre-1.10 string representation. 

I haven't run benchmarks yet, but they should be relatively unchanged, since this is all stuff that gets computed at compile time.