Fix module prefixing for type names (#64)
Through a bit of debugging I learned that `string(Type)` will include the `parentmodule(Type)` as a prefix in some contexts and not others (😢). This affects the hash of types (at least until #55 merges) and therefore affects pretty much any hash value (since they all hash their type). This PR checks for the `parentmodule` prefix in `string(Type)` to make sure that the downstream output is consistent.

I opted to fix this bug only for a new hash version=3 because it is entirely possible that existing uses of the older hash versions depend on the old, broken behavior. Indeed, some of the reference tests I wrote depend on it.