Support Julia 1.11 (hash version=4 only). (#69)
Julia hash versions 1-3 have a number of components that make them unstable when Julia upgrades and in fact, they break sometimes when being compiled under Julia 1.11. (In normal use I seem to be able to call versions 1-3, but the tests fail in various ways that look like they are related to generated functions). We've just built a new hash version that should be much more stable and work on Julia 1.11 (#58), and the maintenance burden of keeping 1-3 working on newer Julia versions isn't something I have the bandwidth for.

Running tests in 1.11 also revealed a subtle bug in buffering, based on a I misunderstanding I had about the `IOBuffer` API. From what I can tell this did not lead to any correctness issues on earlier versions (because I was carefully resetting buffer state to make sure buffer positions lined up with the start of the array), but I was not following public API, and so it broke on 1.11. This was fixed by using `take!` instead of the gymnastics I was going through before.