Fix `Pair` bug (#77)
Transformer for Pair returned a tuple. This means that the types for a `Pair` field can go from `Any` to the concrete type of a value in the `Tuple`. This violates the assumptions of `hoist_type=true`. To fix the problem, `Pair` is now treated as an `UnorderedStruct` rather than as a `DictType`.