stablize type hash of `Base.Fix1`  and `Base.Fix2` for julia 1.12
`Fix1` and `Fix2` got generailzed to `Fix{N}` in Julia 1.12, changing their `nameof` output, which is used by `StableHashTraits` to hash the type of an object. This fixes the string output of `Fix{1}` to `Fix1` and `Fix{2}` to `Fix2` to avoid the regression.

Reviewers: palday

Reviewed By: palday

Pull Request: https://github.com/beacon-biosignals/StableHashTraits.jl/pull/86